AudioFileManager: Fix warning
Patch #63


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@15601 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ src/sound/AudioFileManager.cpp
@@ -513,7 +513,7 @@ AudioFileManager::createRecordingAudioFile(QString projectName, QString instrume
     QString fnm(fileName);
     const QString fpath = m_audioPath + fileName;
     try {
-        aF = new WAVAudioFile( static_cast<const unsigned int>(newId), qstrtostr(fileName), fpath );
+        aF = new WAVAudioFile(newId, qstrtostr(fileName), fpath);
         //aF = new WAVAudioFile(newId, fileName.data(), m_audioPath + qstrtostr(fileName) );
         m_audioFiles.push_back(aF);
         m_recordedAudioFiles.insert(aF);
